fix: improve file search to always start at cwd with upward search
if we search upward starting at the current buffer, but the buffer lies above cwd, we will not find the appropriate taskfile. This has the side effect, that if no file is found and force create is enabled, it will overwrite a currently existing file. This is, because the create logic uses cwd as location.